ci: run ruff, and clear what it found
The linter was configured in pyproject.toml but nothing ran it, so its
findings had been piling up unseen. Seven of them, all small: a mutable
class default, an unparenthesised implicit concatenation, a naked
try/except/pass, a redundant int(round(...)), an unsorted import block
and a missing executable bit.

ruff is pinned, because a new release adds rules and would turn CI red
on a branch that changed nothing.

+++ scripts/build_icons.py
@@ -36,7 +36,7 @@ MARK = (232, 93, 58)         # 朱。読み取り枠
 
 
 def _lerp(a: int, b: int, t: float) -> int:
-    return int(round(a + (b - a) * t))
+    return round(a + (b - a) * t)
 
 
 def _vertical_gradient(size: int, top: tuple, bottom: tuple) -> Image.Image:
